# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Chiron API",
    "version": "0.1.0"
  },
  "paths": {
    "/healthz": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Healthcheck",
        "operationId": "healthcheck_healthz_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Healthcheck Healthz Get"
                }
              }
            }
          }
        }
      }
    },
    "/v1/agent-runs": {
      "get": {
        "tags": [
          "agent-runs"
        ],
        "summary": "List Agent Runs",
        "operationId": "list_agent_runs_v1_agent_runs_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AgentRun"
                  },
                  "type": "array",
                  "title": "Response List Agent Runs V1 Agent Runs Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "agent-runs"
        ],
        "summary": "Create Agent Run",
        "operationId": "create_agent_run_v1_agent_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentRunCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRun"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/experiments/generate": {
      "post": {
        "tags": [
          "experiments"
        ],
        "summary": "Generate Experiment",
        "operationId": "generate_experiment_api_experiments_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExperimentCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/experiments/{experiment_id}": {
      "get": {
        "tags": [
          "experiments"
        ],
        "summary": "Get Experiment",
        "operationId": "get_experiment_api_experiments__experiment_id__get",
        "parameters": [
          {
            "name": "experiment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Experiment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/experiments/{experiment_id}/start-plan": {
      "post": {
        "tags": [
          "experiments"
        ],
        "summary": "Start Plan",
        "operationId": "start_plan_api_experiments__experiment_id__start_plan_post",
        "parameters": [
          {
            "name": "experiment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Experiment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AgentRun": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "agent_name": {
            "type": "string",
            "title": "Agent Name"
          },
          "status": {
            "$ref": "#/components/schemas/AgentRunStatus",
            "default": "queued"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "agent_name"
        ],
        "title": "AgentRun"
      },
      "AgentRunCreateRequest": {
        "properties": {
          "agent_name": {
            "type": "string",
            "minLength": 1,
            "title": "Agent Name"
          },
          "input": {
            "type": "object",
            "title": "Input"
          }
        },
        "type": "object",
        "required": [
          "agent_name"
        ],
        "title": "AgentRunCreateRequest"
      },
      "AgentRunStatus": {
        "type": "string",
        "enum": [
          "queued",
          "running",
          "completed",
          "failed"
        ],
        "title": "AgentRunStatus"
      },
      "ExperimentCreateRequest": {
        "properties": {
          "question": {
            "type": "string",
            "title": "Question"
          }
        },
        "type": "object",
        "required": [
          "question"
        ],
        "title": "ExperimentCreateRequest"
      },
      "ExperimentResponse": {
        "properties": {
          "experiment_id": {
            "type": "string",
            "title": "Experiment Id"
          },
          "question": {
            "type": "string",
            "title": "Question"
          },
          "status": {
            "$ref": "#/components/schemas/ExperimentStatus"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "experiment_id",
          "question",
          "status",
          "created_at"
        ],
        "title": "ExperimentResponse"
      },
      "ExperimentStatus": {
        "type": "string",
        "enum": [
          "running",
          "lqc_completed",
          "planning",
          "completed"
        ],
        "title": "ExperimentStatus"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}